pyproject.toml
[project]
name = "obelix"
version = "0.1.0"
description = "A multi-provider LLM agent framework with tool support and hooks system"
authors = [{ name = "Giulio Surya Lo Verde" }]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.13"

dependencies = [
    "pydantic>=2.12.0",
    "numpy>=2.2.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "tenacity>=9.0.0",
    "networkx>=3.3",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.71.0"]
openai = ["openai>=2.14.0"]
oci = ["oci>=2.164.0"]
ibm = ["ibm_watsonx_ai>=1.4.0"]
ollama = ["ollama>=0.6.0"]
vllm = ["vllm>=0.13.0"]
mcp = ["mcp>=1.25.0"]

all-llm = [
    "anthropic>=0.71.0",
    "openai>=2.14.0",
    "oci>=2.164.0",
    "ibm_watsonx_ai>=1.4.0",
    "ollama>=0.6.0",
    "vllm>=0.13.0",
]
all = [
    "anthropic>=0.71.0",
    "openai>=2.14.0",
    "oci>=2.164.0",
    "ibm_watsonx_ai>=1.4.0",
    "ollama>=0.6.0",
    "vllm>=0.13.0",
    "mcp>=1.25.0",
]

[project.scripts]
obelix = "obelix.main:main"

[project.urls]
Homepage = "https://github.com/GiulioSurya/Obelix"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obelix"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
line-ending = "lf"
